Rename provider hooks to context
